adding key/value to months
