+++ strr-web/components/bcros/form-section/contact-information/ContactInfo.vue
@@ -9,22 +9,12 @@
           />
         </UFormGroup>
         <UFormGroup name="month" class="desktop:pr-[16px] flex-grow mobile:mb-[16px]">
-          <UDropdown
-            :items="months"
+          <USelect
+            v-model="month"
+            :options="getMonths()"
+            option-attribute="key"
             class="w-full"
-            :label="t('create-account.contact-form.month')"
-            :popper="{
-              placement: 'bottom-start',
-            }"
-          >
-            <UInput
-              v-model="month"
-              class="w-full"
-              color="white"
-              trailing-icon="i-heroicons-chevron-down-20-solid"
-              :placeholder="t('create-account.contact-form.month')"
-            />
-          </UDropdown>
+          />
         </UFormGroup>
         <UFormGroup name="birthYear">
           <UInput v-model="year" :placeholder="t('create-account.contact-form.year')" />
@@ -35,20 +25,36 @@
 </template>
 
 <script setup lang="ts">
-import { DropdownItem } from '@nuxt/ui/dist/runtime/types'
-
 const t = useNuxtApp().$i18n.t
 
 const {
-  months,
   dobOptional
 } = defineProps<{
-  months: DropdownItem[][],
   dobOptional?: boolean
 }>()
 
 const day = defineModel<string>('day')
 const month = defineModel<string>('month')
 const year = defineModel<string>('year')
+
+const months: string[] = [
+  t('general.january'),
+  t('general.february'),
+  t('general.march'),
+  t('general.april'),
+  t('general.may'),
+  t('general.june'),
+  t('general.july'),
+  t('general.august'),
+  t('general.september'),
+  t('general.october'),
+  t('general.november'),
+  t('general.december')
+]
+
+const getMonths = (): { key: string, value: string }[] => months.map((month: string, index: number) => ({
+  value: (index + 1).toLocaleString('en-US', { minimumIntegerDigits: 2, useGrouping: false }),
+  key: month
+}))
 
 </script>
