+++ strr-web/interfaces/dropdown-item-with-value-i.ts
@@ -1,0 +1,5 @@
+import { DropdownItem } from "@nuxt/ui/dist/runtime/types/dropdown";
+
+export interface DropdownItemWithValueI extends DropdownItem {
+  value: string | number
+}
